requirements.txt
flask
flask-cors
Pillow
ultralytics
opencv-python
pytesseract
